docs: add Staker exercise prerequisites
I was unable to fuzz the Staker contract with the ` echidna-test . --contract EchidnaTemplate --config config.yaml`. Running this command returns the following errors: 
- `echidna-test: Couldn't compile given file`
- `ERROR:CryticCompile:Error HH12: Trying to use a non-local installation of Hardhat, which is not supported.`
- `FileNotFoundError: [Errno 2] No such file or directory: 'artifacts/build-info'`

Installing the node modules with `npm install` solved the issue.

I know this might be obvious to some people, but it took me a while to figure that one out.